pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spanish-flashcard-builder"
version = "0.1.0"
description = "A tool for building and managing Spanish vocabulary flashcards"
authors = [
    { name = "Bryson Thill" },
]
requires-python = ">=3.11"
dependencies = [
    "requests~=2.31.0",
    "python-dotenv~=1.0.0",
    "openai~=1.12.0",
    "pillow~=10.2.0",
    "spacy~=3.7.2",
    "click~=8.1.7",
    "pyyaml~=6.0.1",
    "genanki~=0.13.1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.9.5",
    "pre-commit==3.6.0",
    "types-requests==2.31.0.20240125",
    "types-pillow==10.2.0.20240111",
    "types-pyyaml==6.0.12.12",
    "types-setuptools==69.1.0.20240125",
    "mypy==1.15.0"
]

[project.scripts]
spanish-flashcard-builder = "spanish_flashcard_builder.cli:main"
sfb = "spanish_flashcard_builder.cli:main"

[tool.setuptools]
packages = ["spanish_flashcard_builder"]

[tool.ruff]
line-length = 88
target-version = "py311"

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
# Enable Pyflakes (`F`), pycodestyle (`E`), and isort (`I`) codes by default.
select = ["E", "F", "I", "B", "W", "C90"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "build/",
    "dist/",
    ".venv/",
    "venv/",
]
